backend/pyproject.toml
[project]
name = "juicypolicy-backend"
version = "0.1.0"
description = "JuicyPolicy Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.32.0",
    "playwright>=1.57.0",
]
